Various fixes for native-component-list
