+++ apps/native-component-list/src/components/FunctionDemo/FunctionDemo.tsx
@@ -54,7 +54,7 @@ type Props = {
   /**
    * Rendering function to render some additional components based on the function's result.
    */
-  renderAdditionalResult?: (result: unknown) => JSX.Element | void;
+  renderAdditionalResult?: (result: any) => JSX.Element | void;
 };
 
 /**
