+++ apps/native-component-list/src/screens/ImagePicker/ImagePickerScreen.tsx
@@ -6,7 +6,7 @@ import FunctionDemo, {
   FunctionDescription,
   FunctionParameter,
 } from '../../components/FunctionDemo';
-import ImageOrVideoPreview from './ImageOrVideoPreview';
+import ImagePickerAssetsList from './ImagePickerAssetsList';
 
 const LAUNCH_PICKER_PARAMETERS: FunctionParameter[] = [
   {
@@ -219,7 +219,7 @@ function ImagePickerScreen() {
           key={idx}
           namespace="ImagePicker"
           {...props}
-          renderAdditionalResult={ImageOrVideoPreview}
+          renderAdditionalResult={ImagePickerAssetsList}
         />
       ))}
     </ScrollView>
